Add start command for linux/mac
