+++ src/Rust/src/commands/mod.rs
@@ -4,6 +4,9 @@ pub use apply::*;
 mod patch;
 pub use patch::*;
 
+mod start;
+pub use start::*;
+
 #[cfg(target_os = "linux")]
 mod apply_linux_impl;
 #[cfg(target_os = "macos")]
@@ -12,9 +15,7 @@ mod apply_osx_impl;
 mod apply_windows_impl;
 
 #[cfg(target_os = "windows")]
-mod start;
-#[cfg(target_os = "windows")]
-pub use start::*;
+mod start_windows_impl;
 
 #[cfg(target_os = "windows")]
 mod install;
